+++ lib/AppInfo/Application.php
@@ -1,23 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 /**
- * @copyright Copyright (c) 2018 Robin Appelman <robin@icewind.nl>
- *
- * @license GNU AGPL version 3 or any later version
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
+ * SPDX-FileCopyrightText: 2018 Robin Appelman <robin@icewind.nl>
+ * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
 namespace OCA\GroupEveryone\AppInfo;
@@ -32,8 +19,10 @@ use OCP\IGroupManager;
 use OCP\User\Events\UserCreatedEvent;
 
 class Application extends App implements IBootstrap {
+	public const APP_ID = 'group_everyone';
+
 	public function __construct(array $urlParams = []) {
-		parent::__construct('group_everyone', $urlParams);
+		parent::__construct(self::APP_ID, $urlParams);
 	}
 
 	public function register(IRegistrationContext $context): void {
@@ -44,7 +33,7 @@ class Application extends App implements IBootstrap {
 		$context->injectFn([$this, 'registerGroupManager']);
 	}
 
-	public function registerGroupManager(IGroupManager $groupManager, GroupBackend $backend) {
+	public function registerGroupManager(IGroupManager $groupManager, GroupBackend $backend): void {
 		$groupManager->addBackend($backend);
 	}
 }